+++ include/gl/io.hpp
@@ -4,5 +4,12 @@
 
 #pragma once
 
+// clang-format off
+
 #include "gl/io/format.hpp"
 #include "gl/io/stream_options_manipulator.hpp"
+
+#include "gl/io/graph_options.hpp"
+#include "gl/io/graph_fio.hpp"
+
+// clang-format on
