+++ include/gl/io/graph_options.hpp
