+++ tests/source/gl/test_alg_coloring.cpp
@@ -2,7 +2,7 @@
 #include "testing/gl/constants.hpp"
 
 #include <gl/algorithm.hpp>
-#include <gl/graph_file_io.hpp>
+#include <gl/io/graph_fio.hpp>
 #include <gl/topology.hpp>
 
 #include <doctest.h>
