+++ tests/source/gl/test_alg_dijkstra.cpp
@@ -5,7 +5,7 @@
 #include "testing/gl/functional.hpp"
 
 #include <gl/algorithm.hpp>
-#include <gl/graph_file_io.hpp>
+#include <gl/io/graph_fio.hpp>
 #include <gl/topology.hpp>
 
 #include <doctest.h>
